Move examples to /penpad/
