+++ demos/gatsby/src/components/EmbeddedPage.tsx
