+++ demos/gatsby/src/components/editor/Editor.tsx
@@ -1,7 +1,7 @@
 import { Page, Penpad as TruePenpad, Specimen } from '@rstacruz/penpad'
 import React, { useMemo, useState } from 'react'
 import useDebounce from '../../utilities/useDebounce'
-import CSS from '../Embedded.module.css'
+import CSS from './Editor.module.css'
 import CodeMirror from './CodeMirror'
 import DemoResult from './DemoResult'
 import useBabel from './useBabel'
