--- demos/gatsby/src/pages/penpad-embedded.js
@@ -1,2 +1,0 @@
-import Embedded from '../components/Embedded'
-export default Embedded
