+++ demos/gatsby/src/pages/penpad/embedded.tsx
@@ -1,0 +1,2 @@
+import EmbeddedPage from '../../components/EmbeddedPage'
+export default EmbeddedPage
